feature(subsystems): re-integrate afk subsystem handling in discord submodule
