+++ subsystems/DiscordRPC/src/main/java/org/terasology/engine/subsystem/discordrpc/DiscordRPCSystem.java
@@ -1,31 +1,22 @@
-/*
- * Copyright 2018 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
 package org.terasology.engine.subsystem.discordrpc;
 
+import org.terasology.entitySystem.entity.EntityRef;
+import org.terasology.entitySystem.event.ReceiveEvent;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterMode;
 import org.terasology.entitySystem.systems.RegisterSystem;
 import org.terasology.game.Game;
+import org.terasology.logic.afk.AfkEvent;
+import org.terasology.logic.players.LocalPlayer;
 import org.terasology.network.NetworkMode;
 import org.terasology.network.NetworkSystem;
 import org.terasology.registry.In;
 
 /**
- * It's a system that runs when a single player or multi player game has been started to process some stuff
- * throw the {@link DiscordRPCSubSystem}.
+ * It's a system that runs when a single player or multi player game has been started to process some stuff throw the
+ * {@link DiscordRPCSubSystem}.
  *
  * @see DiscordRPCSubSystem
  */
@@ -34,6 +25,9 @@ public class DiscordRPCSystem extends BaseComponentSystem {
 
     @In
     private Game game;
+
+    @In
+    private LocalPlayer player;
 
     @In
     private NetworkSystem networkSystem;
@@ -47,6 +41,33 @@ public class DiscordRPCSystem extends BaseComponentSystem {
             mode = "Solo | " + game.getName();
         }
         return mode;
+    }
+
+    @ReceiveEvent
+    public void onAfk(AfkEvent event, EntityRef entityRef) {
+        if (requireConnection() && player.getClientEntity().equals(entityRef)) {
+            return;
+        }
+        if (event.isAfk()) {
+            disableDiscord();
+        } else {
+            enableDiscord();
+        }
+    }
+
+    private boolean requireConnection() {
+        NetworkMode networkMode = networkSystem.getMode();
+        return networkMode != NetworkMode.CLIENT && networkMode != NetworkMode.DEDICATED_SERVER;
+    }
+
+    private void enableDiscord() {
+        DiscordRPCSubSystem.tryToDiscover();
+        DiscordRPCSubSystem.setState("Idle", true);
+    }
+
+    private void disableDiscord() {
+        DiscordRPCSubSystem.tryToDiscover();
+        DiscordRPCSubSystem.setState(getGame(), true);
     }
 
     @Override
